+++ plot/time_series_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/mlange-42/arche-model/model"
 	"github.com/mlange-42/arche-model/system"
 	"github.com/mlange-42/arche-pixel/plot"
+	"github.com/mlange-42/arche-pixel/window"
 	"github.com/mlange-42/arche/ecs"
 )
 
@@ -18,12 +19,15 @@ func ExampleTimeSeries() {
 	m.Tps = 30
 	m.Fps = 0
 
-	// Create a plot.
-	pl := plot.TimeSeries{
-		Observer: &ExampleObserver{},
-	}
-	// Add the plot as UI system.
-	m.AddUISystem(&pl)
+	// Create a time series plot.
+	// See below for the implementation of the RowObserver.
+	m.AddUISystem(&window.Window{
+		Drawers: []window.Drawer{
+			&plot.TimeSeries{
+				Observer: &RowObserver{},
+			},
+		},
+	})
 
 	// Add a termination system that ends the simulation.
 	m.AddSystem(&system.FixedTermination{
@@ -34,17 +38,17 @@ func ExampleTimeSeries() {
 	// Due to the use of the OpenGL UI system, the model must be run via [github.com/faiface/pixel/pixelgl].
 	// Uncomment the next line. It is commented out as the CI has no display device to test the model run.
 
-	//pixelgl.Run(m.Run)
+	// pixelgl.Run(m.Run)
 }
 
-// ExampleObserver to generate random time series.
-type ExampleObserver struct{}
+// RowObserver to generate random time series.
+type RowObserver struct{}
 
-func (o *ExampleObserver) Initialize(w *ecs.World) {}
-func (o *ExampleObserver) Update(w *ecs.World)     {}
-func (o *ExampleObserver) Header(w *ecs.World) []string {
+func (o *RowObserver) Initialize(w *ecs.World) {}
+func (o *RowObserver) Update(w *ecs.World)     {}
+func (o *RowObserver) Header() []string {
 	return []string{"A", "B", "C"}
 }
-func (o *ExampleObserver) Values(w *ecs.World) []float64 {
+func (o *RowObserver) Values(w *ecs.World) []float64 {
 	return []float64{rand.Float64(), rand.Float64() + 1, rand.Float64() + 2}
 }